[FIX] share,base: avoid loading order issues
Move share column to base instead of module share.

The column is used to filter some groups and make them hidden from the res.users
form. However, if a module was loaded before the module share (e.g. depends only
from base), when upgraded, the eventual update of a res.group regenerated the
res.user view (to compute in_group_XX fields). This recomputation was done
before the module share was loaded in the registry (due to the dependency graph)
and the override of get_application_groups was not done, including shared groups
in the user form.

As we can not guarantee that a module is loaded, better to move to base module.
As share depends only of mail and is autoinstall, it was installed on most
instances anyway.

Fixes #6324
Fixes #5820
