+++ addons/share/__init__.py
@@ -20,7 +20,6 @@
 ##############################################################################
 
 import ir_model
-import res_users
 import wizard
 
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
