+++ openerp/addons/base/res/res_users_view.xml
@@ -43,6 +43,8 @@
             <field name="arch" type="xml">
                 <search string="Groups">
                     <field name="name" filter_domain="['|', ('name','ilike',self), ('category_id','ilike',self)]" string="Group"/>
+                    <field name="share"/>
+                    <filter name="no_share" string="Internal Groups" domain="[('share','=',False)]"/>
                 </search>
             </field>
         </record>
@@ -55,10 +57,11 @@
                     <group col="4">
                         <field name="category_id"/>
                         <field name="name"/>
+                        <field name="share"/>
                     </group>
                     <notebook>
                         <page string="Users">
-                            <field name="users"/>
+                            <field name="users" context="{'search_default_no_share':1}"/>
                         </page>
                         <page string="Inherited">
                             <label string="Users added to this group are automatically added in the following groups."/>
@@ -115,6 +118,7 @@
             <field name="type">ir.actions.act_window</field>
             <field name="res_model">res.groups</field>
             <field name="view_type">form</field>
+            <field name="context">{'search_default_no_share': 1}</field>
             <field name="help">A group is a set of functional areas that will be assigned to the user in order to give them access and rights to specific applications and tasks in the system. You can create custom groups or edit the ones existing by default in order to customize the view of the menu that users will be able to see. Whether they can have a read, write, create and delete access right can be managed from here.</field>
         </record>
         <menuitem action="action_res_groups" id="menu_action_res_groups" parent="base.menu_users" groups="base.group_no_one"/>
@@ -232,6 +236,8 @@
                 <search string="Users">
                     <field name="name" filter_domain="['|', '|', ('name','ilike',self), ('login','ilike',self), ('email','ilike',self)]" string="User"/>
                     <field name="company_ids" string="Company" groups="base.group_multi_company"/>
+                    <field name="share"/>
+                    <filter name="no_share" string="Internal Users" domain="[('share','=',False)]"/>
                 </search>
             </field>
         </record>
@@ -251,6 +257,7 @@
             <field name="view_type">form</field>
             <field name="view_id" ref="view_users_tree"/>
             <field name="search_view_id" ref="view_users_search"/>
+            <field name="context">{'search_default_no_share': 1}</field>
             <field name="help">Create and manage users that will connect to the system. Users can be deactivated should there be a period of time during which they will/should not connect to the system. You can assign them groups in order to give them specific access to the applications they need to use in the system.</field>
         </record>
         <record id="action_res_users_view1" model="ir.actions.act_window.view">
